:sparkles: added messenger message format
also removed showUsername and boldUsername as they're unnecessary now
