+++ lib/config/getConfig.js
@@ -12,13 +12,12 @@ module.exports = async configPath => {
   var defaultConfig = {
     messenger: {
       forceLogin: false,
-      showUsername: true,
-      boldUsername: false,
       filter: {
         whitelist: [],
         blacklist: []
       },
-      separateImages: false
+      separateImages: false,
+      format: '*{username}*: {message}',
     },
     discord: {
       renameChannels: true,
