+++ lib/config/getEnvConfig.js
@@ -7,8 +7,6 @@ module.exports = () => {
       username: env.MESSENGER_USERNAME,
       password: env.MESSENGER_PASSWORD,
       forceLogin: yn(env.MESSENGER_FORCE_LOGIN),
-      showUsername: yn(env.MESSENGER_SHOW_USERNAME),
-      boldUsername: yn(env.MESSENGER_BOLD_USERNAME),
       filter: {
         whitelist: (env.MESSENGER_FILTER_WHITELIST || '').split(',').filter(el => el).map(el => el.trim()), // sort out empty elements and trim whitespace
         blacklist: (env.MESSENGER_FILTER_BLACKLIST || '').split(',').filter(el => el).map(el => el.trim()) // sort out empty elements and trim whitespace
