--- lib/discord/handleCustomEmoji.js
@@ -1,9 +1,0 @@
-module.exports = message => {
-  var matches = message.match(/<:[a-zA-Z0-9_]+:[0-9]+>/g)
-  if (!matches) return message
-  for (let match of matches) {
-    var emoji = match.split(':')[1].split(':')[0]
-    message = message.replace(match, `:${emoji}:`)
-  }
-  return message
-}
