+++ lib/discord/index.js
@@ -1,6 +1,5 @@
 module.exports = {
   createMessage: require('./createMessage'),
   getChannel: require('./getChannel'),
-  handleCustomEmoji: require('./handleCustomEmoji'),
   handleMentions: require('./handleMentions')
 }
