+++ lib/listeners/discord.js
@@ -1,7 +1,7 @@
 const log = require('npmlog')
 const sendError = require('../error')
 
-const { handleCustomEmoji } = require('../discord')
+const { createMessage } = require('../messenger')
 
 module.exports = opts => {
   try {
@@ -38,9 +38,7 @@ module.exports = opts => {
     var username = message.member ? (message.member.nickname || message.author.username) : message.author.username
     log.verbose('discordListener: username', username)
     var msg = {
-      body: handleCustomEmoji(
-        config.messenger.showUsername ? (config.messenger.boldUsername ? `*${username}*: ${content}` : `${username}: ${content}`) : content
-      ),
+      body: createMessage(config, username, content),
       url: message.attachments.size > 0 ? message.attachments.first().url : imageUrl
     }
     if (config.messenger.separateImages) msg = [ { body: msg.body }, { url: msg.url } ]
