+++ lib/messenger/index.js
@@ -1,4 +1,5 @@
 module.exports = {
+  createMessage: require('./createMessage'),
   filter: require('./filter'),
   getChannelName: require('./getChannelName'),
   getSender: require('./getSender'),
